src/server/requirements.txt
aiofile
rawpy
python-xmp-toolkit
onnxruntime-gpu
torch
torchvision
torchaudio
pandas
torch-ort
onnx2torch